+++ magda-web-client/src/actions/datasetSearchActions.js
@@ -96,16 +96,16 @@ export function fetchSearchResultsIfNeeded(urlQueryObject: Object): Store {
     };
 }
 
-export function updatePublishers(publishers: Array<Object>): FacetAction {
+export function updateOrganisations(organisations: Array<Object>): FacetAction {
     return {
-        type: actionTypes.UPDATE_PUBLISHERS,
-        items: publishers
+        type: actionTypes.UPDATE_ORGANISATIONS,
+        items: organisations
     };
 }
 
-export function resetPublisher(): FacetAction {
+export function resetOrganisation(): FacetAction {
     return {
-        type: actionTypes.RESET_PUBLISHER
+        type: actionTypes.RESET_ORGANISATION
     };
 }