+++ magda-web-client/src/actions/featuredDatasetsActions.js
@@ -35,7 +35,7 @@ export function fetchFeaturedDatasetsFromRegistry(ids: Array<string>): Object {
                 config.registryApiUrl +
                     `records/${encodeURIComponent(
                         id
-                    )}?aspect=dcat-dataset-strings&optionalAspect=dataset-publisher&optionalAspect=source&dereference=true`
+                    )}?aspect=dcat-dataset-strings&optionalAspect=dataset-organisation&optionalAspect=source&dereference=true`
             ).then(response => response.json())
         );
         Promise.all(fetches)